refactor timely_update_vid_embedding
