+++ hugegraph-llm/src/hugegraph_llm/demo/rag_demo/vector_graph_block.py
@@ -17,6 +17,7 @@
 
 # pylint: disable=E1101
 
+import asyncio
 import gradio as gr
 
 from hugegraph_llm.config import prompt
@@ -28,6 +29,7 @@ from hugegraph_llm.utils.graph_index_utils import (
     import_graph_data,
 )
 from hugegraph_llm.utils.vector_index_utils import clean_vector_index, build_vector_index, get_vector_index_info
+from hugegraph_llm.utils.log import log
 
 def store_prompt(doc, schema, example_prompt):
     # update env variables: doc, schema and example_prompt
@@ -139,3 +141,16 @@ def create_vector_graph_block():
     tab_upload_text.select(fn=on_tab_select, inputs=[input_file, input_text], outputs=[input_file, input_text])
 
     return input_text, input_schema, info_extract_template
+
+async def timely_update_vid_embedding():
+    while True:
+        try:
+            await asyncio.to_thread(update_vid_embedding)
+            log.info("rebuild_vid_index timely executed successfully.")
+        except asyncio.CancelledError as ce:
+            log.info("Periodic task has been cancelled due to: %s", ce)
+            break
+        except Exception as e:
+            log.error("Failed to execute rebuild_vid_index: %s", e, exc_info=True)
+            raise Exception("Failed to execute rebuild_vid_index") from e
+        await asyncio.sleep(3600)
